// src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        Material: {
            readonly id: number;
            category: number;
            readonly category_name: string;
            name: string;
            unit: number;
            readonly unit_code: string;
            /** Is Sheet/Board */
            is_sheet?: boolean;
            /**
             * Length (m)
             * Format: decimal
             */
            length?: string | null;
            /**
             * Width (m)
             * Format: decimal
             */
            width?: string | null;
            /**
             * Thickness (mm)
             * Format: decimal
             */
            thickness?: string | null;
            readonly area_m2: string;
            barcode?: string | null;
            /**
             * Minimum Threshold
             * Format: decimal
             */
            min_threshold?: string;
            /** Format: decimal */
            readonly average_price: string;
            /** Format: decimal */
            readonly total_stock: string;
            /** Format: decimal */
            readonly available_stock: string;
            /** Format: date-time */
            readonly created_at: string;
            /** Format: date-time */
            readonly updated_at: string;
        };
        MaterialCategory: {
            readonly id: number;
            name: string;
            description?: string | null;
            parent?: number | null;
        };
        MaterialReservation: {
            readonly id: number;
            /**
             * Order ID Reference
             * Format: int64
             */
            order_reference_id: number;
            /**
             * Reserved Quantity
             * Format: decimal
             */
            quantity: string;
            status?: components["schemas"]["MaterialReservationStatusEnum"];
            /** Format: date-time */
            readonly created_at: string;
            /** Format: date-time */
            readonly updated_at: string;
            material: number;
        };
        /**
         * @description * `ACTIVE` - Faol (Rezervda)
         *     * `CONSUMED` - Ishlatildi (Chiqim qilindi)
         *     * `CANCELLED` - Bekor qilindi
         * @enum {string}
         */
        MaterialReservationStatusEnum: "ACTIVE" | "CONSUMED" | "CANCELLED";
        MeasurementUnit: {
            readonly id: number;
            name: string;
            code: string;
        };
        Offcut: {
            readonly id: number;
            readonly material_name: string;
            /**
             * Length (m)
             * Format: decimal
             */
            length: string;
            /**
             * Width (m)
             * Format: decimal
             */
            width: string;
            /** Format: int64 */
            quantity?: number;
            status?: components["schemas"]["OffcutStatusEnum"];
            /** Format: int64 */
            original_order_id?: number | null;
            /** Format: date-time */
            readonly created_at: string;
            /** Format: date-time */
            readonly updated_at: string;
            material: number;
        };
// ...
        PaginatedMaterialCategoryList: {
            data: components["schemas"]["MaterialCategory"][];
            pagination: {
                /** @example 123 */
                count: number;
                /**
                 * Format: uri
                 * @example http://api.example.org/accounts/?page=4
                 */
                next?: string | null;
                /**
                 * Format: uri
                 * @example http://api.example.org/accounts/?page=2
                 */
                previous?: string | null;
            };
        };
        PaginatedMaterialList: {
            data: components["schemas"]["Material"][];
            pagination: {
                /** @example 123 */
                count: number;
                /**
                 * Format: uri
                 * @example http://api.example.org/accounts/?page=4
                 */
                next?: string | null;
                /**
                 * Format: uri
                 * @example http://api.example.org/accounts/?page=2
                 */
                previous?: string | null;
            };
        };
        PaginatedMaterialReservationList: {
            data: components["schemas"]["MaterialReservation"][];
            pagination: {
                /** @example 123 */
                count: number;
                /**
                 * Format: uri
                 * @example http://api.example.org/accounts/?page=4
                 */
                next?: string | null;
                /**
                 * Format: uri
                 * @example http://api.example.org/accounts/?page=2
                 */
                previous?: string | null;
            };
        };
        PaginatedMeasurementUnitList: {
            data: components["schemas"]["MeasurementUnit"][];
            pagination: {
                /** @example 123 */
                count: number;
                /**
                 * Format: uri
                 * @example http://api.example.org/accounts/?page=4
                 */
                next?: string | null;
                /**
                 * Format: uri
                 * @example http://api.example.org/accounts/?page=2
                 */
                previous?: string | null;
            };
        };
        PaginatedOffcutList: {
            data: components["schemas"]["Offcut"][];
            pagination: {
                /** @example 123 */
                count: number;
                /**
                 * Format: uri
                 * @example http://api.example.org/accounts/?page=4
                 */
                next?: string | null;
                /**
                 * Format: uri
                 * @example http://api.example.org/accounts/?page=2
                 */
                previous?: string | null;
            };
        };
// ...
        PatchedMaterial: {
            readonly id?: number;
            category?: number;
            readonly category_name?: string;
            name?: string;
            unit?: number;
            readonly unit_code?: string;
            /** Is Sheet/Board */
            is_sheet?: boolean;
            /**
             * Length (m)
             * Format: decimal
             */
            length?: string | null;
            /**
             * Width (m)
             * Format: decimal
             */
            width?: string | null;
            /**
             * Thickness (mm)
             * Format: decimal
             */
            thickness?: string | null;
            readonly area_m2?: string;
            barcode?: string | null;
            /**
             * Minimum Threshold
             * Format: decimal
             */
            min_threshold?: string;
            /** Format: decimal */
            readonly average_price?: string;
            /** Format: decimal */
            readonly total_stock?: string;
            /** Format: decimal */
            readonly available_stock?: string;
            /** Format: date-time */
            readonly created_at?: string;
            /** Format: date-time */
            readonly updated_at?: string;
        };
        PatchedMaterialCategory: {
            readonly id?: number;
            name?: string;
            description?: string | null;
            parent?: number | null;
        };
        PatchedMaterialReservation: {
            readonly id?: number;
            /**
             * Order ID Reference
             * Format: int64
             */
            order_reference_id?: number;
            /**
             * Reserved Quantity
             * Format: decimal
             */
            quantity?: string;
            status?: components["schemas"]["MaterialReservationStatusEnum"];
            /** Format: date-time */
            readonly created_at?: string;
            /** Format: date-time */
            readonly updated_at?: string;
            material?: number;
        };
        PatchedMeasurementUnit: {
            readonly id?: number;
            name?: string;
            code?: string;
        };
        PatchedOffcut: {
            readonly id?: number;
            readonly material_name?: string;
            /**
             * Length (m)
             * Format: decimal
             */
            length?: string;
            /**
             * Width (m)
             * Format: decimal
             */
            width?: string;
            /** Format: int64 */
            quantity?: number;
            status?: components["schemas"]["OffcutStatusEnum"];
            /** Format: int64 */
            original_order_id?: number | null;
            /** Format: date-time */
            readonly created_at?: string;
            /** Format: date-time */
            readonly updated_at?: string;
            material?: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}